+++ server/src/main/java/org/opensearch/action/admin/cluster/stats/TransportClusterStatsAction.java
@@ -130,7 +130,8 @@ public class TransportClusterStatsAction extends TransportNodesAction<
             clusterService.getClusterName(),
             responses,
             failures,
-            state
+            state,
+            request
         );
     }